using uint32_t as loop counter.  Discarding bad time logging strategies.

+++ src/one_neutral/ref_val___________/bench.cpp
@@ -4,12 +4,12 @@
 int main() {
   Setup();
   int e = 0;
-  for (uint64_t i = 0; i < WARMUP_ITERATIONS; ++i) {
+  for (uint32_t i = 0; i < WARMUP_ITERATIONS; ++i) {
     caller(e);
   }
   TimeLogger logger;
   NOP_SLED_HEAD_1;
-  for (uint64_t i = 0; i < ITERATIONS; ++i) {
+  for (uint32_t i = 0; i < ITERATIONS; ++i) {
     caller(e);
   }
   NOP_SLED_TAIL_1;
