+++ src/one_neutral/tls_error_val_____/bench.cpp
@@ -3,12 +3,12 @@
 
 int main() {
   Setup();
-  for (uint64_t i = 0; i < WARMUP_ITERATIONS; ++i) {
+  for (uint32_t i = 0; i < WARMUP_ITERATIONS; ++i) {
     caller();
   }
   TimeLogger logger;
   NOP_SLED_HEAD_1;
-  for (uint64_t i = 0; i < ITERATIONS; ++i) {
+  for (uint32_t i = 0; i < ITERATIONS; ++i) {
     caller();
   }
   NOP_SLED_TAIL_1;
